+++ assets/universe/locale/en-GB/betty.po
@@ -7,7 +7,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Betty VERSION\n"
 "Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: 2026-04-21 21:49+0100\n"
+"POT-Creation-Date: 2026-04-22 22:54+0100\n"
 "PO-Revision-Date: 2026-02-17 18:12+0000\n"
 "Last-Translator: Bart Feenstra <bart@bartfeenstra.com>\n"
 "Language: en_GB\n"
@@ -336,6 +336,9 @@ msgid "Building"
 msgstr "Building"
 
 msgid "Buildings"
+msgstr ""
+
+msgid "Built-in"
 msgstr ""
 
 msgid "Burial"
@@ -1574,6 +1577,17 @@ msgstr ""
 msgid "Serve a generated site"
 msgstr "Serve a generated site"
 
+msgid ""
+"Serve your site using Python's built-in web server. This is for local use"
+" only, and unsafe for publishing your site."
+msgstr ""
+
+msgid "Server"
+msgstr ""
+
+msgid "Servers"
+msgstr ""
+
 #, python-brace-format
 msgid "Serving your site at {url}..."
 msgstr "Serving your site at {url}..."
@@ -1780,6 +1794,9 @@ msgstr ""
 msgid ""
 "The values of the following HTML attributes will automatically be "
 "replaced with the URLs generated from them where possible: {attributes}"
+msgstr ""
+
+msgid "The web server to use."
 msgstr ""
 
 #, python-brace-format
@@ -2683,6 +2700,12 @@ msgstr[1] ""
 #, python-brace-format
 msgid "{count} serializer"
 msgid_plural "{count} serializers"
+msgstr[0] ""
+msgstr[1] ""
+
+#, python-brace-format
+msgid "{count} server"
+msgid_plural "{count} servers"
 msgstr[0] ""
 msgstr[1] ""
