+++ betty/plugins/server/__init__.py
@@ -1,0 +1,3 @@
+"""
+Server plugins.
+"""
