+++ betty/service/plugin/collection/keyed.py
@@ -8,8 +8,10 @@ from typing import final, override
 
 from betty.collection.keyed import KeyedCollection
 from betty.collection.keyed.adapter import KeyedCollectionAdapter
+from betty.collection.keyed.error import ErroringKeyedCollection
 from betty.machine_name import MachineName
 from betty.plugin import PluginDefinition
+from betty.plugin.error import PluginNotFound
 from betty.plugin.resolve import ResolvablePluginId, resolve_plugin_id
 from betty.service.plugin import PluginServiceProvider
 from betty.service.plugin.collection import CollectionPluginServiceManager
@@ -17,6 +19,10 @@ from betty.service.plugin.collection import CollectionPluginServiceManager
 type KeyedPluginCollectionService[PluginDefinitionT: PluginDefinition, ItemT] = (
     KeyedCollection[MachineName, ResolvablePluginId[PluginDefinitionT], ItemT]
 )
+
+
+class _PluginNotFound(PluginNotFound, KeyError):
+    pass
 
 
 class KeyedCollectionPluginServiceManager[
@@ -42,12 +48,19 @@ class KeyedCollectionPluginServiceManager[
     def new_service(
         self, service_provider: ServiceProviderT, /
     ) -> KeyedPluginCollectionService[PluginDefinitionT, GetServiceItemT]:
-        return KeyedCollectionAdapter(
-            {
-                self.resolve_init_plugin_id(plugin): self.new_service_item(
-                    service_provider, plugin
-                )
-                for plugin in self.get_plugins(service_provider)
-            },
-            key_resolver=resolve_plugin_id,
+        return ErroringKeyedCollection(
+            KeyedCollectionAdapter(
+                {
+                    self.resolve_init_plugin_id(plugin): self.new_service_item(
+                        service_provider, plugin
+                    )
+                    for plugin in self.get_plugins(service_provider)
+                },
+                key_resolver=resolve_plugin_id,
+            ),
+            lambda error, key: _PluginNotFound(
+                self.plugin_type,
+                key,
+                map(self.resolve_init_plugin_id, self.get_plugins(service_provider)),
+            ),
         )
