+++ betty/tests/plugins/server/test_builtin.py
@@ -1,0 +1,28 @@
+import requests
+from pytest_mock import MockerFixture
+from requests import Response
+
+from betty.functools import Do
+from betty.plugins.server.builtin import Builtin
+from betty.project import Project
+
+
+class TestBuiltin:
+    async def test__start_stop_and_public_url(
+        self, isolated_project: Project, mocker: MockerFixture
+    ) -> None:
+        mocker.patch("webbrowser.open_new_tab")
+        content = "Hello, and welcome to my site!"
+        isolated_project.www_directory.mkdir(parents=True)
+        with open(
+            isolated_project.www_directory / "index.html", "w", encoding="utf-8"
+        ) as f:
+            f.write(content)
+        async with await Builtin.new(isolated_project) as server:
+
+            def _assert_response(response: Response) -> None:
+                assert response.status_code == 200
+                assert content == response.content.decode("utf-8")
+                assert response.headers["Cache-Control"] == "no-cache"
+
+            await Do(requests.get, server.public_url).until(_assert_response)
